Fix AddPlayerPacket

+++ RoadRunner/network/packets/add_player_packet.hpp
@@ -12,17 +12,19 @@ namespace RoadRunner {
             class AddPlayerPacket {
             public:
                 static const uint8_t packet_id;
-                int client_guid;
+                uint64_t client_guid;
                 RakNet::RakString username;
-                int entity_id;
+                int32_t entity_id;
                 float x;
                 float y;
                 float z;
-                int pitch;
-                int yaw;
-                int item_id = 0;
-                int item_meta = 0;
+                int8_t pitch;
+                int8_t yaw;
+                uint16_t item_id = 0;
+                ;
+                uint16_t item_meta = 0;
                 RoadRunner::network::types::MetadataType metadata;
+
                 bool deserialize_body(RakNet::BitStream *stream);
                 void serialize_body(RakNet::BitStream *stream);
             };
